feat(MainContainer): check authentication during app load
Rationale:
So far, once authenticated, we are redirected to the feeds page, but when
reloading that page, we are back to the authenticate page (because
fetchAndHandleAuthedUser is only ran inside the AuthenticateContainer.).

The authentication status has to persist until logged out, so firebase checks
the status whenever the application loads.  Essentially, that means the
fetchAndHandleAuthedUser functionality is recreated in MainContainer.

Because the authentication validation in MainContainer#componentDidMount is
asynchronous, render() needs to check isFetching. Returning 'null' is a
shortcut, obviously a spinner or something
would be prettier.