fix(MainContainer): ensure home screen displayed after logout
Rationale:
When we are log out, an empty page is returned.

That's because MainContainer#componentDidMount will not do anything when logged
out - in particular, isFetching will remain in the default state true.

So if we are not authenticated, we need to set isFetching to false in
componentDidMount. MainComponent#render will return the <div> with
isAuthed===false, therefore redirect to the home screen.