feat(DuckDetails): start creation of duck details view
DuckDetailsContainer:26 sets isFetching to false to demonstrate
the following:

As of now, refreshing the duck-details page fails: The redux store
will have been reset and have no duck anymore, so the duckId passed
into the DuckContainer will not suffice to retrieve the actual
duck object. When no duck with that duckId is in the store, we
have to fetch it.